requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0  # For environment variables
pydantic
pydantic[email]
sqlalchemy
bcrypt
python-jose